refactor: Introduced IngressConfig to encapsulate Ingress rules and Tls configs
