--- jkube-kit/config/resource/src/test/resources/ingress-rule-config.json
@@ -1,14 +1,0 @@
-{
-  "host": "example.com",
-  "paths": [{
-    "pathType": "ImplementationSpecific",
-    "path": "/path",
-    "serviceName": "service-name",
-    "servicePort": 8080,
-    "resource": {
-      "apiGroup": "group.k8s.io",
-      "kind": "ResourceKind",
-      "name": "resource-name"
-    }
-  }]
-}
