+++ quickstarts/maven/spring-boot/pom.xml
@@ -213,27 +213,29 @@
                 </config>
               </enricher>
               <resources>
-                <ingressTlsConfigs>
-                  <ingressTlsConfig>
-                    <hosts>
-                      https-example.foo.com
-                    </hosts>
-                    <secretName>testsecret-tls</secretName>
-                  </ingressTlsConfig>
-                </ingressTlsConfigs>
-                <ingressRules>
-                  <ingressRule>
-                    <host>hello.example.systems</host>
-                    <paths>
-                      <path>
-                        <pathType>ImplementationSpecific</pathType>
-                        <path>/</path>
-                        <serviceName>${project.artifactId}</serviceName>
-                        <servicePort>8080</servicePort>
-                      </path>
-                    </paths>
-                  </ingressRule>
-                </ingressRules>
+                <ingress>
+                  <ingressTlsConfigs>
+                    <ingressTlsConfig>
+                      <hosts>
+                        https-example.foo.com
+                      </hosts>
+                      <secretName>testsecret-tls</secretName>
+                    </ingressTlsConfig>
+                  </ingressTlsConfigs>
+                  <ingressRules>
+                    <ingressRule>
+                      <host>hello.example.systems</host>
+                      <paths>
+                        <path>
+                          <pathType>ImplementationSpecific</pathType>
+                          <path>/</path>
+                          <serviceName>${project.artifactId}</serviceName>
+                          <servicePort>8080</servicePort>
+                        </path>
+                      </paths>
+                    </ingressRule>
+                  </ingressRules>
+                </ingress>
               </resources>
             </configuration>
           </plugin>
